# admin/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Event Planner API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/mailings/": {
      "post": {
        "tags": [
          "mailings"
        ],
        "summary": "Create a mailing",
        "description": "Create a new mailing.\n\nOnly super administrators may create mailings.  The ``filters`` field\naccepts a JSON object specifying criteria (e.g., event_id,\nis_paid, is_attended) to select recipients.  If\n``scheduled_at`` is provided, the mailing may be scheduled for\nlater execution by an external scheduler; immediate sending can\nbe triggered via the send endpoint.",
        "operationId": "create_mailing_api_v1_mailings__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "mailings"
        ],
        "summary": "List mailings",
        "description": "List mailings.\n\nOnly super administrators may view the list.  Results are paginated.",
        "operationId": "list_mailings_api_v1_mailings__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by 'created_at' or 'scheduled_at'",
              "title": "Sort By"
            },
            "description": "Sort by 'created_at' or 'scheduled_at'"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort order 'asc' or 'desc'",
              "title": "Order"
            },
            "description": "Sort order 'asc' or 'desc'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailingRead"
                  },
                  "title": "Response List Mailings Api V1 Mailings  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mailings/{mailing_id}": {
      "delete": {
        "tags": [
          "mailings"
        ],
        "summary": "Delete a mailing",
        "description": "\u0423\u0434\u0430\u043b\u0438\u0442\u044c \u0440\u0430\u0441\u0441\u044b\u043b\u043a\u0443.\n\nOnly super administrators may delete mailings along with their logs.  Returns\n204 on success; raises 404 if the mailing does not exist.",
        "operationId": "delete_mailing_api_v1_mailings__mailing_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mailing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mailing Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "mailings"
        ],
        "summary": "Get mailing details",
        "description": "Retrieve a mailing by ID.\n\nOnly super administrators may view mailing details.  Raises 404 if the\nmailing does not exist.",
        "operationId": "get_mailing_api_v1_mailings__mailing_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mailing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mailing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "mailings"
        ],
        "summary": "Update a mailing",
        "description": "Update an existing mailing.\n\nOnly super administrators may update mailings.  Any fields omitted from the\nrequest body will be left unchanged.  If the ``messengers`` list is\nprovided, the existing messenger list will be replaced and associated\ntasks will be recreated with the new schedule and channels.",
        "operationId": "update_mailing_api_v1_mailings__mailing_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mailing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mailing Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mailings/{mailing_id}/send": {
      "post": {
        "tags": [
          "mailings"
        ],
        "summary": "Execute a mailing",
        "description": "Send the mailing to all recipients matching its filters.\n\nOnly super administrators may trigger sending.  Returns the number of\nrecipients to whom the message was sent.",
        "operationId": "send_mailing_api_v1_mailings__mailing_id__send_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mailing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mailing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Send Mailing Api V1 Mailings  Mailing Id  Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mailings/{mailing_id}/logs": {
      "get": {
        "tags": [
          "mailings"
        ],
        "summary": "Get mailing logs",
        "description": "Retrieve delivery logs for a mailing.\n\nOnly super administrators may view logs.  Results are paginated.",
        "operationId": "get_logs_api_v1_mailings__mailing_id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mailing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mailing Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailingLogRead"
                  },
                  "title": "Response Get Logs Api V1 Mailings  Mailing Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MailingCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the mailing for internal reference"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Message content to send to users"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters",
            "description": "Criteria to select recipients, e.g. {\"event_id\": 1, \"is_paid\": true, \"is_attended\": true}"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At",
            "description": "When to schedule the mailing; if null, send immediately"
          },
          "messengers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messengers",
            "description": "List of messenger channels to send this mailing to (e.g. ['telegram','vk','max']). If omitted, no tasks will be generated for bots and the mailing will only be stored for record keeping. Supported channels correspond to available bot integrations."
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "MailingCreate",
        "description": "Schema for creating a new mailing."
      },
      "MailingLogRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "mailing_id": {
            "type": "integer",
            "title": "Mailing Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "sent_at": {
            "type": "string",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mailing_id",
          "user_id",
          "status",
          "error_message",
          "sent_at"
        ],
        "title": "MailingLogRead",
        "description": "Schema for reading a single mailing log entry."
      },
      "MailingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "messengers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messengers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_by",
          "title",
          "content",
          "filters",
          "scheduled_at",
          "created_at",
          "messengers"
        ],
        "title": "MailingRead",
        "description": "Schema for reading a mailing."
      },
      "MailingUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "messengers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messengers"
          }
        },
        "type": "object",
        "title": "MailingUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}